+++ src/westpa/westext/adaptvoronoi/adaptVor_driver.py
@@ -155,7 +155,7 @@ class AdaptiveVoronoiDriver:
             else:
                 self.system.bin_mapper = VoronoiBinMapper(self.dfunc, self.centers, dfargs=dfargs, dfkwargs=dfkwargs)
             self.ncenters = self.system.bin_mapper.nbins
-            new_target_counts = np.empty((self.ncenters,), np.int)
+            new_target_counts = np.empty((self.ncenters,), int)
             new_target_counts[...] = self.walk_count
             self.system.bin_target_counts = new_target_counts
         except (ValueError, TypeError) as e:
